requirements.txt
pandas
numpy
scikit-learn
joblib
pytest
ruff
pyarrow
azureml-core==1.56.0
